backend/requirements.txt
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Pydantic==0.8.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pydantic==1.8.2
PyJWT==2.1.0
python-dateutil==2.8.2
python-dotenv==0.19.0
six==1.16.0
SQLAlchemy==1.4.25
typing-extensions==3.10.0.2
waitress==2.0.0
Werkzeug==2.0.1
